feat(rsc-runtime)!: remove structural JSX — AgentBundle/Skill/Script/McpServer/McpApp/Operation elements and defineRscAgentBundle are gone; JSX renders, config and conventions structure (RFC #63)
