+++ packages/rsc-runtime/package.json
@@ -45,7 +45,6 @@
     "typecheck": "tsc -p tsconfig.build.json --noEmit"
   },
   "peerDependencies": {
-    "agent-bundle": "^0.1.0",
     "react": "^19.2.0"
   },
   "dependencies": {
@@ -58,7 +57,6 @@
     "@rslib/core": "0.23.2",
     "@rstest/core": "0.11.10",
     "@types/react": "19.2.18",
-    "agent-bundle": "workspace:*",
     "react": "19.2.8"
   }
 }
