+++ packages/rsc-runtime/src/plugin.ts
@@ -12,14 +12,3 @@ export type {
   RscOperationDefinition,
   RscOperationInput,
 } from './operation.js';
-export { defineRscAgentBundle } from './plugin-definition.js';
-export type { RscAgentBundleApplication } from './plugin-definition.js';
-export { AgentBundle, McpApp, McpServer, Operation, Script, Skill } from './plugin-elements.js';
-export type {
-  AgentBundleProps,
-  McpAppProps,
-  McpServerProps,
-  OperationProps,
-  ScriptProps,
-  SkillProps,
-} from './plugin-elements.js';
